moved RDS from timing to data
